refactor : 변수 타입 변경

+++ src/main/java/com/example/toanyone/domain/menu/controller/MenuController.java
@@ -18,7 +18,7 @@ public class MenuController {
 
     @PostMapping("/{storeId}/menus")
     public ResponseEntity<MenuDto.Response> createMenu(
-            @PathVariable Integer storeId,
+            @PathVariable Long storeId,
             @Valid @RequestBody MenuDto.Request requestDto) {
 
         MenuDto.Response response = menuService.createMenu(
